Use templates instead for version specific metadata handling
We all make some designing mistakes sometimes.
I thought that making a function in a header file that uses the same struct names in its definition, then inlining that and including it in a .cpp file for the specific metadata version would work, but sadly it didn't. Worked out better in my head.
For anyone wondering, inline is not going to 100% replace the function body into whereever you call it. It's more of a request for the compiler, which it can ignore. And during dev builds and under a few other circumstances, it will.
tl;dr: use templates if you want to make a function for handling something for different versions easily